+++ packages/devtools/test/docker/compose.test.ts
@@ -1,0 +1,60 @@
+import { serializeDockerComposeSpec } from '@/docker/compose'
+import { spawnSync } from 'child_process'
+import { rm, writeFile } from 'fs/promises'
+import { join } from 'path'
+
+describe('docker/compose', () => {
+    describe('serializeDockerComposeSpec', () => {
+        const SPEC_FILE_PATH = join(__dirname, 'docker-compose.yaml')
+
+        const validateSpec = () => spawnSync('docker', ['compose', '-f', SPEC_FILE_PATH, 'config'])
+
+        afterEach(async () => {
+            await rm(SPEC_FILE_PATH, { force: true })
+        })
+
+        it('should create a valid compose spec if called with no services', async () => {
+            const spec = serializeDockerComposeSpec({
+                version: '3.9',
+            })
+
+            await writeFile(SPEC_FILE_PATH, spec)
+
+            expect(validateSpec().status).toBe(0)
+
+            expect(spec).toMatchSnapshot()
+        })
+
+        it('should create a valid compose spec if called with basic services', async () => {
+            const spec = serializeDockerComposeSpec({
+                version: '3.9',
+                services: {
+                    redis: {
+                        image: 'redis:latest',
+                    },
+                    codebase: {
+                        build: '.',
+                        command: 'pnpm build',
+                        volumes: ['./packages:/app/packages'],
+                    },
+                    postgres: {
+                        image: 'postgres',
+                        volumes: ['pgdata:/var/lib/postgresql/data'],
+                    },
+                },
+                volumes: {
+                    pgdata: null,
+                    somevolume: {
+                        name: 'somevolumename',
+                    },
+                },
+            })
+
+            await writeFile(SPEC_FILE_PATH, spec)
+
+            expect(validateSpec().status).toBe(0)
+
+            expect(spec).toMatchSnapshot()
+        })
+    })
+})