+++ examples/oft/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/oft-example
 
+## 0.1.8
+
+### Patch Changes
+
+- 54cf16e: Silence bigint-buffer warning
+
 ## 0.1.7
 
 ### Patch Changes