+++ packages/create-lz-oapp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-lz-oapp
 
+## 0.1.16
+
+### Patch Changes
+
+- 54cf16e: Silence bigint-buffer warning
+
 ## 0.1.15
 
 ### Patch Changes