+++ packages/devtools-evm/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @layerzerolabs/devtools-evm
 
+## 0.2.9
+
+### Patch Changes
+
+- cc83f96: Add fp-ts to dependencies
+- Updated dependencies [40c45c1]
+- Updated dependencies [40c45c1]
+- Updated dependencies [4292d81]
+  - @layerzerolabs/devtools@0.2.8
+
+## 0.2.8
+
+### Patch Changes
+
+- 54cf16e: Silence bigint-buffer warning
+- Updated dependencies [54cf16e]
+- Updated dependencies [2b9ae6a]
+- Updated dependencies [2b9ae6a]
+  - @layerzerolabs/devtools@0.2.7
+  - @layerzerolabs/io-devtools@0.1.5
+
 ## 0.2.7
 
 ### Patch Changes