+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/devtools
 
+## 0.2.8
+
+### Patch Changes
+
+- 40c45c1: Add tapError utility to simplify async error logging
+- 40c45c1: Add mapError utility
+- 4292d81: Add docker compose spec generation capabilities
+
+## 0.2.7
+
+### Patch Changes
+
+- 54cf16e: Silence bigint-buffer warning
+- 2b9ae6a: Sign transactions for different chains in parallel
+- 2b9ae6a: Add groupTransactionsByEid utility
+- Updated dependencies [54cf16e]
+  - @layerzerolabs/io-devtools@0.1.5
+
 ## 0.2.6
 
 ### Patch Changes