+++ packages/verify-contract/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @layerzerolabs/verify-contract
 
+## 1.1.12
+
+### Patch Changes
+
+- f9572ae: Updated urls for metis, mantle, scroll
+- eaea2e3: Added default scan urls for fraxtal and mode
+
+## 1.1.11
+
+### Patch Changes
+
+- 54cf16e: Silence bigint-buffer warning
+
 ## 1.1.10
 
 ### Patch Changes